separate SemesterPicker component
