+++ src/components/Problems/Problems.module.scss
@@ -56,26 +56,3 @@
   font-weight: bold;
   margin-bottom: 1rem;
 }
-
-
-.notPublished {
-  background-color: #EEE;
-}
-
-
-.menu {
-  position: fixed;
-  top: 5rem;
-  right: 0px;
-  width: 25vw;
-  height: 5rem;
-  z-index: 1000;
-  user-select: none;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.problemSubmission {
-  border: 1px solid red;
-}
