+++ src/components/Problems/Problems.tsx
@@ -11,31 +11,9 @@ import {useSeminarInfo} from '@/utils/useSeminarInfo'
 
 import {Latex} from '../Latex/Latex'
 import {Discussion} from './Discussion'
-import {Dropdown, DropdownOption} from './Dropdown'
 import styles from './Problems.module.scss'
+import {SemesterList, SemesterPicker} from './SemesterPicker'
 import {UploadProblemForm} from './UploadProblemForm'
-
-interface SeriesList {
-  id: number
-  order: number
-  deadline: string
-  complete: boolean
-  frozen_results: string | null
-  semester: number
-}
-
-interface SemesterList {
-  id: number
-  year: number
-  school_year: string
-  season_code: number
-  start: string
-  end: string
-  frozen_results: boolean
-  competition: number
-  late_tags: string[]
-  series_set: SeriesList[]
-}
 
 interface Problem {
   id: number
@@ -258,7 +236,7 @@ export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
             <CircularProgress color="inherit" />
           </div>
         )}
-        <Menu semesterList={semesterList} selectedSeriesId={seriesId} />
+        <SemesterPicker semesterList={semesterList} selectedSeriesId={seriesId} />
         {problems.map((problem) => (
           <Problem
             key={problem.id}
@@ -302,46 +280,6 @@ export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
         )}
       </div>
     </>
-  )
-}
-
-const Menu: FC<{semesterList: SemesterList[]; selectedSeriesId: number}> = ({semesterList, selectedSeriesId}) => {
-  const {seminar} = useSeminarInfo()
-
-  let selectedSemesterId = -1
-
-  const dropdownSemesterList = semesterList.map((semester) => {
-    return {
-      id: semester.id,
-      text: `${semester.year}. Ročník - ${semester.season_code === 0 ? 'zimný' : 'letný'} semester`,
-      link: `/${seminar}/zadania/${semester.year}/${semester.season_code === 0 ? 'zima' : 'leto'}/`,
-    }
-  })
-
-  let dropdownSeriesList: DropdownOption[] = []
-
-  for (let i = 0; i < semesterList.length; i++) {
-    for (let j = 0; j < semesterList[i].series_set.length; j++) {
-      if (semesterList[i].series_set[j].id === selectedSeriesId) {
-        selectedSemesterId = semesterList[i].id
-        dropdownSeriesList = semesterList[i].series_set.map((series) => {
-          return {
-            id: series.id,
-            text: `${series.order}. séria`,
-            link: `/${seminar}/zadania/${semesterList[i].year}/${semesterList[i].season_code === 0 ? 'zima' : 'leto'}/${
-              series.order
-            }/`,
-          }
-        })
-      }
-    }
-  }
-
-  return (
-    <div className={styles.menu}>
-      <Dropdown title={'Séria'} selectedId={selectedSeriesId} options={dropdownSeriesList} />
-      <Dropdown title={'Semester'} selectedId={selectedSemesterId} options={dropdownSemesterList} />
-    </div>
   )
 }
