+++ src/components/Problems/SemesterPicker.module.scss
@@ -1,0 +1,12 @@
+.menu {
+  position: fixed;
+  top: 5rem;
+  right: 0px;
+  width: 25vw;
+  height: 5rem;
+  z-index: 1000;
+  user-select: none;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
