+++ src/components/Problems/SemesterPicker.module.scss.d.ts
@@ -1,0 +1,9 @@
+export type Styles = {
+  menu: string
+}
+
+export type ClassNames = keyof Styles
+
+declare const styles: Styles
+
+export default styles
